Crate gear_core_processor

Gear message processor.

Modules

common

Common structures for processing.

configs

Configurations.

Structs

Ext

Structure providing externalities for running host functions.

PreparedMessageExecutionContext

Checked parameters for message execution across processing runs.

ProcessExecutionContext

Checked parameters for message execution across processing runs.

ProcessorContext

Processor context.

Enums

PrechargeResult

Defines result variants of the function precharge.

PrepareResult

Defines result variants of the function prepare.

ProcessorError

Ext’s error

Constants

ERR_EXIT_CODE

Error exit code.

RE_INIT_EXIT_CODE

A try to init again initialized, existing program.

UNAVAILABLE_DEST_EXIT_CODE

Destination isn’t available for the message.

Traits

ProcessorExt

Trait to which ext must have to work in processor wasm executor. Currently used only for lazy-pages support.

Functions

calculate_gas_for_code

Calculates gas amount required to charge for code loading.

calculate_gas_for_program

Calculates gas amount required to charge for program loading.

execute_wasm

Execute wasm with dispatch and return dispatch result.

handle_journal

Handle some journal records passing them to the journal handler.

precharge

Charge a message for program data beforehand.

prepare

Prepares environment for the execution of a program. Checks if there is a required export and tries to charge for code and memory pages. Returns either PreparedMessageExecutionContext for process or an array of journal notes. See PrepareResult for details.

process

Process program & dispatch for it and return journal for updates.