Enum gear_core_processor::common::ExecutionErrorReason

pub enum ExecutionErrorReason {

    Memory(MemoryError),
    Backend(String),
    Ext(TrapExplanation),
    NonExecutable,
    NotLastPage,
    LoadMemoryGasExceeded,
    LoadMemoryBlockGasExceeded,
    GrowMemoryGasExceeded,
    GrowMemoryBlockGasExceeded,
    InitialMemoryGasExceeded,
    InitialMemoryBlockGasExceeded,
    InsufficientMemorySize,
    PageNoData,
    PageIsNotAllocated(PageNumber),
    LazyPagesInitFailed(String),
    InitialMemoryReadFailed(PageNumber, MemoryError),
    InitialDataWriteFailed(PageNumber, MemoryError),
    OutOfRent,
    InitialPagesContainsDataInLazyPagesMode,

}

Reason of execution error

Variants

Memory(MemoryError)

Memory error

Backend(String)

Backend error

Ext(TrapExplanation)

Ext error

NonExecutable

Not executable actor.

NotLastPage

Program’s max page is not last page in wasm page

LoadMemoryGasExceeded

Not enough gas to load memory

LoadMemoryBlockGasExceeded

Not enough gas in block to load memory

GrowMemoryGasExceeded

Not enough gas to grow memory size

GrowMemoryBlockGasExceeded

Not enough gas in block to grow memory size

InitialMemoryGasExceeded

Not enough gas for initial memory handling

InitialMemoryBlockGasExceeded

Not enough gas in block for initial memory handling

InsufficientMemorySize

Mem size less then static pages num

PageNoData

Changed page has no data in initial pages

PageIsNotAllocated(PageNumber)

Page with data is not allocated for program

LazyPagesInitFailed(String)

Lazy pages init failed for current program.

InitialMemoryReadFailed(PageNumber, MemoryError)

Cannot read initial memory data from wasm memory.

InitialDataWriteFailed(PageNumber, MemoryError)

Cannot write initial data to wasm memory.

OutOfRent

Message killed from storage as out of rent.

InitialPagesContainsDataInLazyPagesMode

Initial pages data must be empty when execute with lazy pages