Enum gear_core_processor::common::JournalNote

pub enum JournalNote {
    MessageDispatched(DispatchOutcome),
    GasBurned {
        message_id: MessageId,
        amount: u64,
    },
    ExitDispatch {
        id_exited: ProgramId,
        value_destination: ProgramId,
    },
    MessageConsumed(MessageId),
    SendDispatch {
        message_id: MessageId,
        dispatch: Dispatch,
    },
    WaitDispatch(StoredDispatch),
    WakeMessage {
        message_id: MessageId,
        program_id: ProgramId,
        awakening_id: MessageId,
    },
    UpdatePage {
        program_id: ProgramId,
        page_number: PageNumber,
        data: Option<Vec<u8>>,
    },
    SendValue {
        from: ProgramId,
        to: Option<ProgramId>,
        value: u128,
    },
    StoreNewPrograms {
        code_hash: CodeId,
        candidates: Vec<(ProgramId, MessageId)>,
    },
}

Journal record for the state update.

Variants

MessageDispatched(DispatchOutcome)

Tuple Fields

0: DispatchOutcome

Message was successfully dispatched.

GasBurned

Fields

message_id: MessageId

Message id in which gas was burned.

amount: u64

Amount of gas burned.

Some gas was burned.

ExitDispatch

Fields

id_exited: ProgramId

Id of the program called exit.

value_destination: ProgramId

Address where all remaining value of the program should be transferred to.

Exit the program.

MessageConsumed(MessageId)

Tuple Fields

0: MessageId

Message was handled and no longer exists.

This should be the last update involving this message id.

SendDispatch

Fields

message_id: MessageId

Message id of the message that generated this message.

dispatch: Dispatch

New message with entry point that was generated.

Message was generated.

WaitDispatch(StoredDispatch)

Tuple Fields

0: StoredDispatch

Put this dispatch in the wait list.

WakeMessage

Fields

message_id: MessageId

Message which has initiated wake.

program_id: ProgramId

Program which has initiated wake.

awakening_id: MessageId

Message that should be wokoen.

Wake particular message.

UpdatePage

Fields

program_id: ProgramId

Program that owns the page.

page_number: PageNumber

Number of the page.

data: Option<Vec<u8>>

New data of the page.

Updates data in case of Some(data) or deletes the page

Update page.

SendValue

Fields

from: ProgramId

Value sender

to: Option<ProgramId>

Value beneficiary,

value: u128

Value amount

Send value

StoreNewPrograms

Fields

code_hash: CodeId

Code hash used to create new programs with ids in candidates field

candidates: Vec<(ProgramId, MessageId)>

Collection of program candidate ids and their init message ids.

Store programs requested by user to be initialized later