Struct gear_core_processor::ProcessorContext

pub struct ProcessorContext {

    pub gas_counter: GasCounter,
    pub gas_allowance_counter: GasAllowanceCounter,
    pub value_counter: ValueCounter,
    pub allocations_context: AllocationsContext,
    pub message_context: MessageContext,
    pub block_info: BlockInfo,
    pub config: AllocationsConfig,
    pub existential_deposit: u128,
    pub origin: ProgramId,
    pub program_id: ProgramId,
    pub program_candidates_data: BTreeMap<CodeId, Vec<(ProgramId, MessageId)>>,
    pub host_fn_weights: HostFnWeights,
    pub forbidden_funcs: BTreeSet<&'static str>,
    pub mailbox_threshold: u64,

}

Processor context.

Fields
gas_counter: GasCounter

Gas counter.

gas_allowance_counter: GasAllowanceCounter

Gas allowance counter.

value_counter: ValueCounter

Value counter.

allocations_context: AllocationsContext

Allocations context.

message_context: MessageContext

Message context.

block_info: BlockInfo

Block info.

config: AllocationsConfig

Allocations config.

existential_deposit: u128

Account existential deposit

origin: ProgramId

Communication origin

program_id: ProgramId

Current program id

program_candidates_data: BTreeMap<CodeId, Vec<(ProgramId, MessageId)>>

Map of code hashes to program ids of future programs, which are planned to be initialized with the corresponding code (with the same code hash).

host_fn_weights: HostFnWeights

Weights of host functions.

forbidden_funcs: BTreeSet<&'static str>

Functions forbidden to be called.

mailbox_threshold: u64

Mailbox threshold