Trait gear_core_processor::ProcessorExt

pub trait ProcessorExt {
    type Error: Display;

    fn new(context: ProcessorContext) -> Self;
    fn is_lazy_pages_enabled() -> bool;
    fn check_lazy_pages_consistent_state() -> bool;
    fn lazy_pages_protect_and_init_info(
        mem: &impl Memory,
        prog_id: ProgramId
    ) -> Result<(), Self::Error>;
    fn lazy_pages_post_execution_actions(
        mem: &impl Memory,
        memory_pages: &mut BTreeMap<PageNumber, PageBuf>
    ) -> Result<(), Self::Error>;
}

Trait to which ext must have to work in processor wasm executor. Currently used only for lazy-pages support.

Required Associated Types

type Error: Display

An error issues in processor

Required Methods

fn new(context: ProcessorContext) -> Self

Create new

fn is_lazy_pages_enabled() -> bool

Returns whether this extension works with lazy pages

fn check_lazy_pages_consistent_state() -> bool

If extension support lazy pages, then checks that environment for lazy pages is initialized.

fn lazy_pages_protect_and_init_info(
mem: &impl Memory,
prog_id: ProgramId
) -> Result<(), Self::Error>

Protect and save storage keys for pages which has no data

fn lazy_pages_post_execution_actions(
mem: &impl Memory,
memory_pages: &mut BTreeMap<PageNumber, PageBuf>
) -> Result<(), Self::Error>

Lazy pages contract post execution actions