Crate gear_core_processor

Gear message processor.

Modules

common

Common structures for processing.

configs

Configurations.

informational

Informational functions for core-processor and executor.

Structs

ContextChargedForCode

The instance returned by precharge_for_code. Existence of the instance means that corresponding counters were successfully charged for fetching the binary code from storage.

ContextChargedForInstrumentation

The instance returned by precharge_for_instrumentation. Existence of the instance means that corresponding counters were successfully charged for reinstrumentation of the code.

ContextChargedForProgram

Struct with dispatch and counters charged for program data.

Ext

Structure providing externalities for running host functions.

ProcessExecutionContext

Checked parameters for message execution across processing runs.

ProcessorContext

Processor context.

SystemReservationContext

System reservation context.

Enums

AllocExtError

Ext’s memory management (calls to allocate and free) error.

FallibleExtError

Fallible API error.

SuccessfulDispatchResultKind

Possible variants of the DispatchResult if the latter contains value.

UnrecoverableExtError

Infallible API error.

Traits

ProcessorExternalities

Trait to which ext must have to work in processor wasm executor. Currently used only for lazy-pages support.

Functions

handle_journal

Handle some journal records passing them to the journal handler.

precharge_for_allocations

Precharge for allocations obtaining from storage.

precharge_for_code

Charge a message for the program binary code beforehand.

precharge_for_code_length

Charge a message for fetching the actual length of the binary code from a storage. The updated value of binary code length should be kept in standalone storage. The caller has to call this function to charge gas-counters accordingly before fetching the value.

precharge_for_instrumentation

Charge a message for instrumentation of the binary code beforehand.

precharge_for_module_instantiation

Charge a message for program memory and module instantiation beforehand.

precharge_for_program

Charge a message for program data beforehand.

process

Process program & dispatch for it and return journal for updates.

process_allowance_exceed

Helper function for journal creation if the block gas allowance has been exceeded.

process_code_not_exists

Helper function for journal creation in code not exists case.

process_execution_error

Helper function for journal creation in trap/error case.

process_failed_init

Helper function for journal creation in program failed init case.

process_program_exited

Helper function for journal creation in program exited case.

process_reinstrumentation_error

Helper function for journal creation in case of re-instrumentation error.

process_success

Helper function for journal creation in success case

process_uninitialized

Helper function for journal creation in program uninitialized case.