Crate gear_core_processor

Gear message processor.

Re-exports

pub use precharge::*;

Modules

common

Common structures for processing.

configs

Configurations.

informational

Informational functions for core-processor and executor.

precharge

Precharge module.

Structs

Ext

Structure providing externalities for running host functions.

ProcessExecutionContext

Checked parameters for message execution across processing runs.

ProcessorContext

Processor context.

SystemReservationContext

System reservation context.

Enums

AllocExtError

Ext’s memory management (calls to allocate and free) error.

FallibleExtError

Fallible API error.

UnrecoverableExtError

Infallible API error.

Traits

ProcessorExternalities

Trait to which ext must have to work in processor wasm executor. Currently used only for lazy-pages support.

Functions

handle_journal

Handle some journal records passing them to the journal handler.

process

Process program & dispatch for it and return journal for updates.

process_allowance_exceed

Helper function for journal creation if the block gas allowance has been exceeded.

process_code_not_exists

Helper function for journal creation in code not exists case.

process_execution_error

Helper function for journal creation in trap/error case.

process_failed_init

Helper function for journal creation in program failed init case.

process_instrumentation_failed

Helper function for journal creation in case of instrumentation failure.

process_program_exited

Helper function for journal creation in program exited case.

process_reinstrumentation_error

Helper function for journal creation in case of re-instrumentation error.

process_success

Helper function for journal creation in success case

process_uninitialized

Helper function for journal creation in program uninitialized case.