Enum gear_core_processor::common::JournalNote

pub enum JournalNote {

    MessageDispatched {
        message_id: MessageId,
        source: ProgramId,
        outcome: DispatchOutcome,
    },
    GasBurned {
        message_id: MessageId,
        amount: u64,
    },
    ExitDispatch {
        id_exited: ProgramId,
        value_destination: ProgramId,
    },
    MessageConsumed(MessageId),
    SendDispatch {
        message_id: MessageId,
        dispatch: Dispatch,
        delay: u32,
        reservation: Option<ReservationId>,
    },
    WaitDispatch {
        dispatch: StoredDispatch,
        duration: Option<u32>,
        waited_type: MessageWaitedType,
    },
    WakeMessage {
        message_id: MessageId,
        program_id: ProgramId,
        awakening_id: MessageId,
        delay: u32,
    },
    UpdatePage {
        program_id: ProgramId,
        page_number: GearPage,
        data: PageBuf,
    },
    UpdateAllocations {
        program_id: ProgramId,
        allocations: BTreeSet<WasmPage>,
    },
    SendValue {
        from: ProgramId,
        to: Option<ProgramId>,
        value: u128,
    },
    StoreNewPrograms {
        code_id: CodeId,
        candidates: Vec<(MessageId, ProgramId)>,
    },
    StopProcessing {
        dispatch: StoredDispatch,
        gas_burned: u64,
    },
    ReserveGas {
        message_id: MessageId,
        reservation_id: ReservationId,
        program_id: ProgramId,
        amount: u64,
        duration: u32,
    },
    UnreserveGas {
        reservation_id: ReservationId,
        program_id: ProgramId,
        expiration: u32,
    },
    UpdateGasReservations {
        program_id: ProgramId,
        reserver: GasReserver,
    },
    SystemReserveGas {
        message_id: MessageId,
        amount: u64,
    },
    SystemUnreserveGas {
        message_id: MessageId,
    },
    SendSignal {
        message_id: MessageId,
        destination: ProgramId,
        err: SimpleSignalError,
    },

}

Journal record for the state update.

Variants

MessageDispatched

Fields

message_id: MessageId

Message id of dispatched message.

source: ProgramId

Source of the dispatched message.

outcome: DispatchOutcome

Outcome of the processing.

Message was successfully dispatched.

GasBurned

Fields

message_id: MessageId

Message id in which gas was burned.

amount: u64

Amount of gas burned.

Some gas was burned.

ExitDispatch

Fields

id_exited: ProgramId

Id of the program called exit.

value_destination: ProgramId

Address where all remaining value of the program should be transferred to.

Exit the program.

MessageConsumed(MessageId)

Message was handled and no longer exists.

This should be the last update involving this message id.

SendDispatch

Fields

message_id: MessageId

Message id of the message that generated this message.

dispatch: Dispatch

New message with entry point that was generated.

delay: u32

Amount of blocks to wait before sending.

reservation: Option<ReservationId>

Whether use supply from reservation or current message.

Message was generated.

WaitDispatch

Fields

dispatch: StoredDispatch

Stored dispatch to be inserted into Waitlist.

duration: Option<u32>

Expected duration of holding.

waited_type: MessageWaitedType

If this message is waiting for its reincarnation.

Put this dispatch in the wait list.

WakeMessage

Fields

message_id: MessageId

Message which has initiated wake.

program_id: ProgramId

Program which has initiated wake.

awakening_id: MessageId

Message that should be woken.

delay: u32

Amount of blocks to wait before waking.

Wake particular message.

UpdatePage

Fields

program_id: ProgramId

Program that owns the page.

page_number: GearPage

Number of the page.

data: PageBuf

New data of the page.

Update page.

UpdateAllocations

Fields

program_id: ProgramId

Program id.

allocations: BTreeSet<WasmPage>

New allocations set for the program.

Update allocations set note. And also removes data for pages which is not in allocations set now.

SendValue

Fields

from: ProgramId

Value sender

to: Option<ProgramId>

Value beneficiary,

value: u128

Value amount

Send value

StoreNewPrograms

Fields

code_id: CodeId

Code hash used to create new programs with ids in candidates field

candidates: Vec<(MessageId, ProgramId)>

Collection of program candidate ids and their init message ids.

Store programs requested by user to be initialized later

StopProcessing

Fields

dispatch: StoredDispatch

Pushes StoredDispatch back to the top of the queue.

gas_burned: u64

Decreases gas allowance by that amount, burned for processing try.

Stop processing queue.

ReserveGas

Fields

message_id: MessageId

Message from which gas is reserved.

reservation_id: ReservationId

Reservation ID

program_id: ProgramId

Program which contains reservation.

amount: u64

Amount of reserved gas.

duration: u32

How many blocks reservation will live.

Reserve gas.

UnreserveGas

Fields

reservation_id: ReservationId

Reservation ID

program_id: ProgramId

Program which contains reservation.

expiration: u32

Block number until reservation will live.

Unreserve gas.

UpdateGasReservations

Fields

program_id: ProgramId

Program whose map will be updated.

reserver: GasReserver

Map with reservations.

Update gas reservation map in program.

SystemReserveGas

Fields

message_id: MessageId

Message ID which system reservation will be made from.

amount: u64

Amount of reserved gas.

Do system reservation.

SystemUnreserveGas

Fields

message_id: MessageId

Message ID which system reservation was made from.

Do system unreservation in case it is created but not used.

SendSignal

Fields

message_id: MessageId

Message ID which system reservation was made from.

destination: ProgramId

Program ID which signal will be sent to.

err: SimpleSignalError

Simple signal error.

Send signal.