gear_core_processor

Module precharge

Precharge module.

Structs

ContextCharged

Context charged gas for the program execution.

ForAllocations

ZST for the context that charged for allocations.

ForCodeMetadata

ZST for the context that charged for code metadata.

ForInstrumentedCode

ZST for the context that charged for instrumented code.

ForModuleInstantiation

ZST for the context that charged for module instantiation.

ForNothing

ZST for the context that charged nothing.

ForOriginalCode

ZST for the context that charged for original code.

ForProgram

ZST for the context that charged for program data.

Enums

PreChargeGasOperation

Operation related to gas charging.

Type Aliases

PrechargeResult

Defines result variants of the precharge functions.