Struct gear_core_processor::Ext

pub struct Ext {

    pub gas_counter: GasCounter,
    pub gas_allowance_counter: GasAllowanceCounter,
    pub value_counter: ValueCounter,
    pub allocations_context: AllocationsContext,
    pub message_context: MessageContext,
    pub block_info: BlockInfo,
    pub config: AllocationsConfig,
    pub existential_deposit: u128,
    pub error_explanation: Option<ExtError>,
    pub exit_argument: Option<ProgramId>,
    pub origin: ProgramId,
    pub program_id: ProgramId,
    pub program_candidates_data: BTreeMap<CodeId, Vec<(ProgramId, MessageId)>>,
    pub host_fn_weights: HostFnWeights,

}

Structure providing externalities for running host functions.

Fields
gas_counter: GasCounter

Gas counter.

gas_allowance_counter: GasAllowanceCounter

Gas allowance counter.

value_counter: ValueCounter

Value counter.

allocations_context: AllocationsContext

Allocations context.

message_context: MessageContext

Message context.

block_info: BlockInfo

Block info.

config: AllocationsConfig

Allocations config.

existential_deposit: u128

Account existential deposit

error_explanation: Option<ExtError>

Any guest code panic explanation, if available.

exit_argument: Option<ProgramId>

Contains argument to the exit if it was called.

origin: ProgramId

Communication origin

program_id: ProgramId

Current program id

program_candidates_data: BTreeMap<CodeId, Vec<(ProgramId, MessageId)>>

Map of code hashes to program ids of future programs, which are planned to be initialized with the corresponding code (with the same code hash).

host_fn_weights: HostFnWeights

Weights of host functions.

Implementations

impl Ext

pub fn return_and_store_err<T>(
&mut self,
result: Result<T, ExtError>
) -> Result<T, ExtError>

Return result and store error info in field