Trait gear_core_processor::ProcessorExt

pub trait ProcessorExt {
    fn new(
        gas_counter: GasCounter, 
        gas_allowance_counter: GasAllowanceCounter, 
        value_counter: ValueCounter, 
        allocations_context: AllocationsContext, 
        message_context: MessageContext, 
        block_info: BlockInfo, 
        config: AllocationsConfig, 
        existential_deposit: u128, 
        error_explanation: Option<&'static str>, 
        exit_argument: Option<ProgramId>, 
        origin: ProgramId, 
        program_id: ProgramId, 
        program_candidates_data: BTreeMap<CodeId, Vec<(ProgramId, MessageId)>>, 
        host_fn_weights: HostFnWeights
    ) -> Self;

    fn try_to_enable_lazy_pages(
        &mut self, 
        program_id: ProgramId, 
        memory_pages: &mut BTreeMap<PageNumber, Option<Box<PageBuf>>>
    ) -> Result<bool, &'static str>;

    fn protect_pages_and_init_info(
        memory_pages: &BTreeMap<PageNumber, Option<Box<PageBuf>>>, 
        prog_id: ProgramId, 
        wasm_mem_begin_addr: u64
    ) -> Result<(), &'static str>;

    fn post_execution_actions(
        memory_pages: &mut BTreeMap<PageNumber, Option<Box<PageBuf>>>, 
        wasm_mem_begin_addr: u64
    ) -> Result<(), &'static str>;

    fn remove_lazy_pages_prot(mem_addr: u64) -> Result<(), &'static str>;

    fn protect_lazy_pages_and_update_wasm_mem_addr(
        old_mem_addr: u64, 
        new_mem_addr: u64
    ) -> Result<(), &'static str>;
}

Trait to which ext must have to work in processor wasm executor. Currently used only for lazy-pages support.

Required methods

fn new(
gas_counter: GasCounter,
gas_allowance_counter: GasAllowanceCounter,
value_counter: ValueCounter,
allocations_context: AllocationsContext,
message_context: MessageContext,
block_info: BlockInfo,
config: AllocationsConfig,
existential_deposit: u128,
error_explanation: Option<&'static str>,
exit_argument: Option<ProgramId>,
origin: ProgramId,
program_id: ProgramId,
program_candidates_data: BTreeMap<CodeId, Vec<(ProgramId, MessageId)>>,
host_fn_weights: HostFnWeights
) -> Self

Create new

fn try_to_enable_lazy_pages(
&mut self,
program_id: ProgramId,
memory_pages: &mut BTreeMap<PageNumber, Option<Box<PageBuf>>>
) -> Result<bool, &'static str>

Try to enable and initialize lazy pages env

fn protect_pages_and_init_info(
memory_pages: &BTreeMap<PageNumber, Option<Box<PageBuf>>>,
prog_id: ProgramId,
wasm_mem_begin_addr: u64
) -> Result<(), &'static str>

Protect and save storage keys for pages which has no data

fn post_execution_actions(
memory_pages: &mut BTreeMap<PageNumber, Option<Box<PageBuf>>>,
wasm_mem_begin_addr: u64
) -> Result<(), &'static str>

Lazy pages contract post execution actions

fn remove_lazy_pages_prot(mem_addr: u64) -> Result<(), &'static str>

Remove lazy-pages protection, returns wasm memory begin addr

fn protect_lazy_pages_and_update_wasm_mem_addr(
old_mem_addr: u64,
new_mem_addr: u64
) -> Result<(), &'static str>

Protect lazy-pages and set new wasm mem addr if it has been changed